+++ cmd/commander/commander.go
@@ -53,14 +53,23 @@ Tests are always executed in alphabetical order.
 
 Examples:
 
+Directory test:
+commander test --dir /your/dir/
+
+Stdin test:
+cat commander.yaml | commander test -
+
+HTTP test:
+commander test https://your-url/commander_test.yaml
+
 Filtering tests:
-test commander.yaml --filter="my test"
+commander test commander.yaml --filter="my test"
 
 Multiple filters:
-test commander.yaml --filter=filter1 --filter=filter2
+commander test commander.yaml --filter=filter1 --filter=filter2
 
 Regex filters:
-test commander.yaml --filter="^filter1$"
+commander test commander.yaml --filter="^filter1$"
 `,
 		ArgsUsage: "[file] [--filter]",
 		Flags: []cli.Flag{