+++ pkg/app/test_command.go
@@ -1,6 +1,7 @@
 package app
 
 import (
+	"bufio"
 	"fmt"
 	"io/ioutil"
 	"log"
@@ -38,6 +39,10 @@ func TestCommand(testPath string, ctx TestCommandContext) error {
 		fmt.Println("Starting test against directory: " + testPath + "...")
 		fmt.Println("")
 		result, err = testDir(testPath, ctx.Filters)
+	case testPath == "-":
+		fmt.Println("Starting test from stdin...")
+		fmt.Println("")
+		result, err = testStdin(ctx.Filters)
 	case isURL(testPath):
 		fmt.Println("Starting test from " + testPath + "...")
 		fmt.Println("")
@@ -119,6 +124,23 @@ func convergeResults(result runtime.Result, new runtime.Result) runtime.Result {
 	result.Duration += new.Duration
 
 	return result
+}
+
+func testStdin(filters runtime.Filters) (runtime.Result, error) {
+	f, err := os.Stdin.Stat()
+	if err != nil {
+		return runtime.Result{}, err
+	}
+
+	if (f.Mode() & os.ModeCharDevice) != 0 {
+		return runtime.Result{}, fmt.Errorf("Error: when testing from stdin the command is intended to work with pipes")
+	}
+
+	r := bufio.NewReader(os.Stdin)
+	content, err := ioutil.ReadAll(r)
+	s := suite.ParseYAML(content, "")
+
+	return execute(s, filters)
 }
 
 func execute(s suite.Suite, filters runtime.Filters) (runtime.Result, error) {